requirements.txt
python-dotenv
pydantic
pandas
pypdf
beautifulsoup4
langchain
langchain-openai
langchain-community
langchain-experimental
langgraph
fastapi